feat: makes nil an option as dark or light themes
If the user wants "not to set" (set to nil), it is now possible for
the variables `auto-dark-dark-theme` and `auto-dark-light-theme`.

This way you can set both to nil and make usage of only the detection
mechanism and its hooks `auto-dark-light-mode-hook` and
`auto-dark-dark-mode-hook`.

You can also set only one of the themes to nil to use "default Emacs
appearance".