Better errors when determining detection method
Rather than just saying that we _couldn’t_ determine the detection method, this
indicates why and makes some suggestions. Probably even longer errors with
concrete steps to remedy each case would be helpful.

This includes an alternative approach to the first bullet point in #61 –
changing where we suggest that `auto-dark-allow-osascript` should be
enabled (although this does not change the error that suggests it in the wrong
place).