Improve interaction with `custom-safe-themes`
This adds some documentation around what you can expect with
`custom-safe-themes`.

It also makes a couple changes to the code so that prompts caused by
`custom-safe-themes` are more likely to happen when customizing
`auto-dark-themes` and less likely to happen during Emacs’ initialization or
when switching between light and dark modes.

And, finally, it tries to ensure that the current themes are enabled immediately
after customizing `auto-dark-themes`, rather than waiting for the next mode
switch.