DBus: Let "no preference" be light again
The value `0` in the `color-scheme` spec[1] means "No preference". The
way `auto-dark-mode` historically has interpreted "No preference" is as
light mode. This was true both in `auto-dark--current-mode-dbus` and
`auto-dark--register-dbus-listener` up until 3bacb8f3998a.

In 3bacb8f3998a this logic changed in `auto-dark--current-mode-dbus` but
was *kept* in `auto-dark--register-dbus-listener`. This means that I now
need to flip the light- and dark mode switch in the OS twice once I've
started Emacs to make the theme in Emacs be in sync with the OS.

Revert the logic change in `auto-dark--current-mode-dbus` introduced in
3bacb8f3998a to make sure it stays in sync with
`auto-dark--register-dbus-listener`.

1: https://flatpak.github.io/xdg-desktop-portal/docs/doc-org.freedesktop.portal.Settings.html